Multiplayer & Relative Turns
Multiplayer : thanks to Roman for adding multiplayer support on the
server : socket var « idPlayer » refer to the right instance and is
used to fetch the keys in the JSON « keys ». Also, all buttons keys are
grouped in an unique socket « butt » with a value and a state (up or
down).
Relative turns : The socket for rotation « rowOrientation » call
convertAngle(rowRotation) to convert the angle of the phone in a value
between 0 and 1. MinAngle define the minimum angle when the turn begin
(higher value helps not turning in strait lines) and maxAngle define
the max steer angle. This computed value is stored in
« CurrentOrientation »
Meanwhile TurnHandler run a statement : if full turn, turnKeyDown, if
relative press:  set up a timeChunk, start pressing key, calculate the
time of pressing, and release key. It’s basically 2 nested Timeout
function !Then it call callBackTH which run everything again (like an
infinite loop).
This one was tricky, but it works :D